+++ src/main/kotlin/com/pestphp/pest/PestSettings.kt
@@ -10,6 +10,12 @@ import com.pestphp.pest.parser.PestConfigurationFileParser
 @State(name = "PestSettings", storages = [Storage("pest.xml")])
 class PestSettings : PersistentStateComponent<PestSettings> {
     var pestFilePath = "tests/Pest.php"
+    var preferredTestFlavor =  TestFlavor.IT
+
+    enum class TestFlavor {
+        IT,
+        TEST
+    }
 
     override fun getState(): PestSettings {
         return this